+++ src/components/tabs/__spec__.js
@@ -53,9 +53,9 @@ describe('Tabs', () => {
   describe('componentWillMount', () => {
     describe('initial selected tab', () => {
       describe('when passed as props', () => {
-        it('uses the prop as initialSelectedTabId and takes precendent over hash', () => {
+        it('uses the prop as selectedTabId and takes precendent over hash', () => {
           instance = TestUtils.renderIntoDocument(
-            <Tabs initialSelectedTabId='uniqueid2'>
+            <Tabs selectedTabId='uniqueid2'>
               <Tab title='Tab Title 1' tabId='uniqueid1'>
                 <Textbox name='foo'/>
                 <Textbox name='bar'/>
@@ -118,6 +118,51 @@ describe('Tabs', () => {
           expect(instanceWithNull.state.selectedTabId).toEqual('uniqueid1');
         });
       });
+    });
+  });
+
+  describe('Change in tab prop', () => {
+    let instance, tabs;
+    beforeEach(() => {
+      let TestParent = React.createFactory(React.createClass({
+        getInitialState() {
+          return { selectedTabId: "uniqueid2", align: 'left' };
+        },
+        render() {
+          return (
+            <Tabs { ...this.state }>
+              <Tab title='Tab Title 1' tabId='uniqueid1'>
+                <Textbox name='foo'/>
+                <Textbox name='bar'/>
+              </Tab>
+              <Tab title='Tab Title 2' tabId='uniqueid2'>
+                <Textbox name='baz'/>
+                <Textbox name='bax'/>
+              </Tab>
+            </Tabs>
+          );
+        }
+      }));
+      instance = TestUtils.renderIntoDocument(TestParent());
+      tabs = TestUtils.scryRenderedComponentsWithType(instance, Tab);
+      expect(tabs[0].props.className).toEqual('hidden');
+      expect(tabs[1].props.className).not.toEqual('hidden');
+    });
+
+    it('changes tab to current prop', () => {
+      instance.setState({
+        selectedTabId: "uniqueid1"
+      });
+      expect(tabs[1].props.className).toEqual('hidden');
+      expect(tabs[0].props.className).not.toEqual('hidden');
+    });
+
+    it('change in other tab', () => {
+      instance.setState({
+        align: 'right'
+      });
+      expect(tabs[0].props.className).toEqual('hidden');
+      expect(tabs[1].props.className).not.toEqual('hidden');
     });
   });
 
@@ -207,13 +252,13 @@ describe('Tabs', () => {
       let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[1];
       expect(secondTab.classList.contains('headerClass2')).toEqual(true);
     });
-    
+
     it('adds the className included in the props as an array to the tab', () => {
       let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[2];
       expect(secondTab.classList.contains('headerClass3')).toEqual(true);
       expect(secondTab.classList.contains('headerClass4')).toEqual(true);
     });
-    
+
     it('does not addsa selected class to the header', () => {
       let secondTab = TestUtils.scryRenderedDOMComponentsWithTag(instance, 'li')[1];
       expect(secondTab.classList.contains('carbon-tabs__headers__header--selected')).toEqual(false);